PPLUS-1142: Console commands integrator logic / review fixes
