+++ docs/manifest-types.md
@@ -48,19 +48,19 @@ This manifest type adds Plugin to desired place (by defining exact method) of th
                 "prepend-arguments": [
                     {
                         "value": "",
-                        "is_literal": false
+                        "is_literal": true
                     }
                 ],
                 "append-arguments": [
                     {
                         "value": "",
-                        "is_literal": false
+                        "is_literal": true
                     }
                 ],
                 "constructor-arguments": [
                     {
                         "value": "",
-                        "is_literal": false
+                        "is_literal": true
                     }
                 ]
             }
