+++ src/Builder/Visitor/AddPluginToPluginCollectionVisitor.php
@@ -11,14 +11,12 @@ namespace SprykerSdk\Integrator\Builder\Visitor;
 
 use PhpParser\BuilderFactory;
 use PhpParser\Node;
-use PhpParser\Node\Arg;
 use PhpParser\Node\Expr\MethodCall;
 use PhpParser\Node\Expr\New_;
 use PhpParser\Node\Stmt\Expression;
 use PhpParser\NodeFinder;
 use PhpParser\NodeVisitorAbstract;
-use SprykerSdk\Integrator\Builder\ArgumentBuilder\ArgumentBuilderInterface;
-use SprykerSdk\Integrator\Helper\ClassHelper;
+use SprykerSdk\Integrator\Builder\Visitor\ArgumentBuilder\ArgumentBuilderInterface;
 use SprykerSdk\Integrator\Transfer\ClassMetadataTransfer;
 
 class AddPluginToPluginCollectionVisitor extends NodeVisitorAbstract
@@ -34,13 +32,13 @@ class AddPluginToPluginCollectionVisitor extends NodeVisitorAbstract
     protected $classMetadataTransfer;
 
     /**
-     * @var \SprykerSdk\Integrator\Builder\ArgumentBuilder\ArgumentBuilderInterface
+     * @var \SprykerSdk\Integrator\Builder\Visitor\ArgumentBuilder\ArgumentBuilderInterface
      */
     protected $argumentBuilder;
 
     /**
      * @param \SprykerSdk\Integrator\Transfer\ClassMetadataTransfer $classMetadataTransfer
-     * @param \SprykerSdk\Integrator\Builder\ArgumentBuilder\ArgumentBuilderInterface $argumentBuilder
+     * @param \SprykerSdk\Integrator\Builder\Visitor\ArgumentBuilder\ArgumentBuilderInterface $argumentBuilder
      */
     public function __construct(ClassMetadataTransfer $classMetadataTransfer, ArgumentBuilderInterface $argumentBuilder)
     {
@@ -104,7 +102,7 @@ class AddPluginToPluginCollectionVisitor extends NodeVisitorAbstract
             }
 
             if ($addPluginCallCount) {
-                $arguments = $this->createAddPluginArguments();
+                $arguments = $this->argumentBuilder->createAddPluginArguments($this->classMetadataTransfer);
                 $newMethodCall = (new BuilderFactory())
                     ->methodCall($addPluginCalls[0]->var, $addPluginCalls[0]->name, $arguments);
 
@@ -115,47 +113,5 @@ class AddPluginToPluginCollectionVisitor extends NodeVisitorAbstract
         }
 
         return $node;
-    }
-
-    /**
-     * @return array<\PhpParser\Node\Arg>
-     */
-    protected function createAddPluginArguments(): array
-    {
-        $args = [];
-        $builderFactory = new BuilderFactory();
-
-        $constructorArgumentValues = [];
-        if ($this->classMetadataTransfer->getConstructorArguments()->count()) {
-            $constructorArgumentValues = $this->argumentBuilder->getArguments(
-                $this->classMetadataTransfer->getConstructorArguments()->getArrayCopy(),
-            );
-        }
-
-        if ($this->classMetadataTransfer->getPrependArguments()->count()) {
-            $prependArgumentValues = $this->argumentBuilder->getArguments(
-                $this->classMetadataTransfer->getPrependArguments()->getArrayCopy(),
-            );
-
-            $args = array_merge($args, $prependArgumentValues);
-        }
-
-        $mainArgument = new Arg(
-            $builderFactory->new(
-                (new ClassHelper())->getShortClassName($this->classMetadataTransfer->getSourceOrFail()),
-                $builderFactory->args($constructorArgumentValues),
-            ),
-        );
-        $args = array_merge($args, $builderFactory->args([$mainArgument]));
-
-        if ($this->classMetadataTransfer->getAppendArguments()->count()) {
-            $appendArgumentValues = $this->argumentBuilder->getArguments(
-                $this->classMetadataTransfer->getAppendArguments()->getArrayCopy(),
-            );
-
-            $args = array_merge($args, $appendArgumentValues);
-        }
-
-        return $args;
     }
 }
